requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
langchain>=0.1.0
llama-index>=0.9.0
requests>=2.31.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
sqlite3